:recycle: [#1579] -- break up DataContainer for more general purpose usage
This was 'owned by' the logic evaluation, but the abstractions are
useful in other places too, so it's broken up into a general purpose
module and a specific subclass.
